requirements.txt
streamlit
PyPDF2
langchain
chromadb
tiktoken
sentence-transformers
huggingface_hub
langchain-community
python-dotenv
langchain-huggingface
langchain-chroma